refactor(windows): extract shared Windows temp launch utilities
Extract duplicated Windows logic from claude_temp_launch.rs and
codex_temp_launch.rs into a new windows_temp_launch.rs module:

- is_cmd_shim, arg_requires_cmd_quote
- quote_windows_arg, quote_windows_arg_for_cmd
- build_windows_command_line, build_env_block_with_override
- ScopedConsoleCtrlHandler, Job
- spawn_suspended_createprocessw, wait_for_child
- restrict_to_owner, create_secret_temp_file

Also add AppError constructors for Windows Job Object failures.

Add an automated Windows smoke test covering:
- spawn suspended child via CreateProcessW
- Job Object creation and assignment
- ResumeThread + wait + exit code propagation
